django-chatapp/requirements.txt
astroid==2.0.1
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
docopt==0.6.2
idna==2.7
isort==4.3.4
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
pbr==4.2.0
pipreqs==0.4.9
Pygments==2.2.0
pylint==2.0.0
pylint-django==2.0
pylint-plugin-utils==0.4
pymodm==0.4.0
pymongo==3.7.1
pytz==2018.5
redgreenunittest==0.1.1
requests==2.19.1
six==1.11.0
urllib3==1.23
wrapt==1.10.11
yarg==0.1.9
